refactor: move SelectInput componet to PoolOperations
